Rework and enable rename feature support
Rework the existing code for the rename feature to only detect renames
for the previous version to current; detecting renames for several
versions back increases computational time significantly, but for just a
minor benefit. The common case is that files are renamed (or renamed and
have mostly the same content) *once in a while*, not all the time.
Calculating renames for several versions back is mostly useful when
files are changing all the time.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>